Act2 16/01/2020

+++ sisper/m_inclusiones/a_tdocumentario/g_bandeja.php
@@ -81,53 +81,62 @@ if(accesocon($cone,$_SESSION['identi'],17)){
                         }
                         mysqli_free_result($cn);
 
-                        $q="INSERT INTO doc (Numero, Ano, Siglas, FechaDoc, idTipoDoc, asunto, folios, remitenteext, destinatarioext, deporigenext, depdestinoext, remitenteint, destinatarioint, deporigenint, depdestinoint, numdoc, fecregistro, regpor, cargo) VALUES ($num, '$ano', $sig, '$fecdoc', $tipdoc, $asu, $fol, $perem, $pedes, $derem, $dedes, $pirem, $pides, $direm, $dides, $nu, NOW(), $idem, $car);";
-                        if(mysqli_query($cone, $q)){
-                            $iddo=mysqli_insert_id($cone);
 
-                            $pmp=mysqli_query($cone, "SELECT pm.idtdpersonalmp FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.idtdmesapartes=$mpd AND pm.estado=1 AND mp.estado=1;");
-                            if(mysqli_num_rows($pmp)>0){
-                                $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idEmpleado, idtdmesapartes, asignador, mpasignador, estado) VALUES ($iddo, 2, NOW(), $idem, $mpd, $idem, $mpd, 1);";
-                                if(mysqli_query($cone, $q)){
-                                    $r['m']=mensajesu("Listo, documento registrado y recibido.<br> N° Seguimiento:<b> $nu-$ano</b>");
-                                    $r['e']=true;
-                                }else{
-                                    if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddo;")){
-                                        $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
+                        $cnu=mysqli_query($cone, "SELECT idDoc FROM doc WHERE numdoc=$nu AND Ano='$ano';");
+                        if(mysqli_num_rows($cnu)>0){
+                            $r['m']=mensajewa("Error, # de seguimiento ya registrado. Vuelva a intentarlo para generar un uno nuevo.");
+                        }else{
+
+                            $q="INSERT INTO doc (Numero, Ano, Siglas, FechaDoc, idTipoDoc, asunto, folios, remitenteext, destinatarioext, deporigenext, depdestinoext, remitenteint, destinatarioint, deporigenint, depdestinoint, numdoc, fecregistro, regpor, cargo) VALUES ($num, '$ano', $sig, '$fecdoc', $tipdoc, $asu, $fol, $perem, $pedes, $derem, $dedes, $pirem, $pides, $direm, $dides, $nu, NOW(), $idem, $car);";
+                            if(mysqli_query($cone, $q)){
+                                $iddo=mysqli_insert_id($cone);
+
+                                $pmp=mysqli_query($cone, "SELECT pm.idtdpersonalmp FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.idtdmesapartes=$mpd AND pm.estado=1 AND mp.estado=1;");
+                                if(mysqli_num_rows($pmp)>0){
+                                    $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idEmpleado, idtdmesapartes, asignador, mpasignador, estado) VALUES ($iddo, 2, NOW(), $idem, $mpd, $idem, $mpd, 1);";
+                                    if(mysqli_query($cone, $q)){
+                                        $r['m']=mensajesu("Listo, documento registrado y recibido.<br> N° Seguimiento:<b> $nu-$ano</b>");
+                                        $r['e']=true;
                                     }else{
-                                        $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
+                                        if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddo;")){
+                                            $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
+                                        }else{
+                                            $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
+                                        }
                                     }
+
+                                }else{
+
+                                    //validamos si deriva de una mesa de partes
+                                    $cmmp=mysqli_query($cone, "SELECT mp.idtdmesapartes FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.estado=1 AND mp.estado=1;");
+                                    if($rmmp=mysqli_fetch_assoc($cmmp)){
+                                        $mmp=vacio($rmmp['idtdmesapartes']);
+                                        $dep=vacio("");
+                                    }else{
+                                        $mmp=vacio("");
+                                        $dep=vacio(iddependenciae($cone, $_SESSION['identi']));
+                                    }
+                                    mysqli_free_result($cmmp);
+                                    $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idtdmesapartes, asignador, depasignador, mpasignador, estado) VALUES ($iddo, 3, NOW(), $mpd, $idem, $dep, $mmp, 1);";
+                                    if(mysqli_query($cone, $q)){
+                                        $r['m']=mensajesu("Listo, documento registrado y derivado.<br> N° Seguimiento:<b> $nu-$ano</b>");
+                                        $r['e']=true;
+                                    }else{
+                                        if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddo;")){
+                                            $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
+                                        }else{
+                                            $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
+                                        }
+                                    }
+
                                 }
 
                             }else{
-
-                                //validamos si deriva de una mesa de partes
-                                $cmmp=mysqli_query($cone, "SELECT mp.idtdmesapartes FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.estado=1 AND mp.estado=1;");
-                                if($rmmp=mysqli_fetch_assoc($cmmp)){
-                                    $mmp=vacio($rmmp['idtdmesapartes']);
-                                    $dep=vacio("");
-                                }else{
-                                    $mmp=vacio("");
-                                    $dep=vacio(iddependenciae($cone, $_SESSION['identi']));
-                                }
-                                mysqli_free_result($cmmp);
-                                $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idtdmesapartes, asignador, depasignador, mpasignador, estado) VALUES ($iddo, 3, NOW(), $mpd, $idem, $dep, $mmp, 1);";
-                                if(mysqli_query($cone, $q)){
-                                    $r['m']=mensajesu("Listo, documento registrado y derivado.<br> N° Seguimiento:<b> $nu-$ano</b>");
-                                    $r['e']=true;
-                                }else{
-                                    if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddo;")){
-                                        $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
-                                    }else{
-                                        $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
-                                    }
-                                }
-
+                                $r['m']=mensajewa("Error al registrar, vuelva a intentarlo.<br> $q");
                             }
 
-                        }else{
-                            $r['m']=mensajewa("Error al registrar, vuelva a intentarlo.<br> $q");
                         }
+                        mysqli_free_result($cnu);
                 }
 
             }else{
@@ -207,38 +216,44 @@ if(accesocon($cone,$_SESSION['identi'],17)){
                         }
                         mysqli_free_result($cn);
 
-                        $q="INSERT INTO doc (Numero, Ano, Siglas, FechaDoc, idTipoDoc, asunto, folios, remitenteext, destinatarioext, deporigenext, depdestinoext, remitenteint, destinatarioint, deporigenint, depdestinoint, numdoc, fecregistro, regpor, cargo) VALUES ($num, '$ano', $sig, '$fecdoc', $tipdoc, $asu, $fol, $perem, $pedes, $derem, $dedes, $pirem, $pides, $direm, $dides, $nu, NOW(), $idem, $car);";
-                        if(mysqli_query($cone, $q)){
-                            $iddo=mysqli_insert_id($cone);
+                        $cnu=mysqli_query($cone, "SELECT idDoc FROM doc WHERE numdoc=$nu AND Ano='$ano';");
+                        if(mysqli_num_rows($cnu)>0){
+                            $r['m']=mensajewa("Error, # de seguimiento ya registrado. Vuelva a intentarlo para generar un uno nuevo.");
+                        }else{
 
-                            $cmp=mysqli_query($cone, "SELECT pm.idtdmesapartes FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.estado=1 AND mp.estado=1;");
-                            if($rmp=mysqli_fetch_assoc($cmp)){
-                                $mp=$rmp['idtdmesapartes'];
-                                $dep=iddependenciae($cone, $ped);
+                            $q="INSERT INTO doc (Numero, Ano, Siglas, FechaDoc, idTipoDoc, asunto, folios, remitenteext, destinatarioext, deporigenext, depdestinoext, remitenteint, destinatarioint, deporigenint, depdestinoint, numdoc, fecregistro, regpor, cargo) VALUES ($num, '$ano', $sig, '$fecdoc', $tipdoc, $asu, $fol, $perem, $pedes, $derem, $dedes, $pirem, $pides, $direm, $dides, $nu, NOW(), $idem, $car);";
+                            if(mysqli_query($cone, $q)){
+                                $iddo=mysqli_insert_id($cone);
 
-                                $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idEmpleado, idDependencia, asignador, mpasignador, pnotificar, estado) VALUES ($iddo, 3, NOW(), $ped, $dep, $idem, $mp, 1, 1);";
+                                $cmp=mysqli_query($cone, "SELECT pm.idtdmesapartes FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.estado=1 AND mp.estado=1;");
+                                if($rmp=mysqli_fetch_assoc($cmp)){
+                                    $mp=$rmp['idtdmesapartes'];
+                                    $dep=iddependenciae($cone, $ped);
 
-                                if(mysqli_query($cone, $q)){
-                                    $r['m']=mensajesu("Listo, documento registrado y asignado.<br> N° Seguimiento:<b> $nu-$ano</b>");
-                                    $r['e']=true;
+                                    $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idEmpleado, idDependencia, asignador, mpasignador, pnotificar, estado) VALUES ($iddo, 3, NOW(), $ped, $dep, $idem, $mp, 1, 1);";
+
+                                    if(mysqli_query($cone, $q)){
+                                        $r['m']=mensajesu("Listo, documento registrado y asignado.<br> N° Seguimiento:<b> $nu-$ano</b>");
+                                        $r['e']=true;
+                                    }else{
+                                        if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddo;")){
+                                            $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
+                                        }else{
+                                            $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
+                                        }
+                                    }
+
                                 }else{
                                     if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddo;")){
-                                        $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
+                                        $r['m']=mensajewa("No pertenece a ninguna mesa de partes.");
                                     }else{
                                         $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
                                     }
                                 }
 
                             }else{
-                                if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddo;")){
-                                    $r['m']=mensajewa("No pertenece a ninguna mesa de partes.");
-                                }else{
-                                    $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
-                                }
+                                $r['m']=mensajewa("Error al registrar, vuelva a intentarlo.<br> $q");
                             }
-
-                        }else{
-                            $r['m']=mensajewa("Error al registrar, vuelva a intentarlo.<br> $q");
                         }
                 }
 
@@ -937,37 +952,42 @@ if(accesocon($cone,$_SESSION['identi'],17)){
                     }
                     mysqli_free_result($cn);
 
-                    //ingresamos doc
-                    if(mysqli_query($cone, "INSERT INTO doc (Numero, Ano, Siglas, FechaDoc, idTipoDoc, folios, remitenteext, destinatarioext, deporigenext, depdestinoext, cargo, remitenteint, destinatarioint, deporigenint, depdestinoint, numdoc, asunto, fecregistro, regpor, idDocRel) VALUES ($numero, '$ano', $siglas, '$fechadoc', $idtipodoc, 1, $remitenteext, $destinatarioext, $deporigenext, $depdestinoext, 1, $remitenteint, $destinatarioint, $deporigenint, $depdestinoint, $nu, '$asunto', NOW(), $idem, $idd);")){
-                        $iddn=mysqli_insert_id($cone);
-
-                        //obtrenemos la mesa de partes
-                        $cmmp=mysqli_query($cone, "SELECT mp.idtdmesapartes FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.estado=1 AND mp.estado=1;");
-                        if($rmmp=mysqli_fetch_assoc($cmmp)){
-                            $mmp=vacio($rmmp['idtdmesapartes']);
-                            $dep=vacio("");
-                        }else{
-                            $mmp=vacio("");
-                            $dep=vacio(iddependenciae($cone, $_SESSION['identi']));
-                        }
-                        mysqli_free_result($cmmp);
-
-                        $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idEmpleado, idtdmesapartes, idDependencia, asignador, mpasignador, depasignador, estado) VALUES ($iddn, 2, NOW(), $idem, $mmp, $dep, $idem, $mmp, $dep, 1);";
-                        if(mysqli_query($cone, $q)){
-                            $r['m']=mensajesu("Listo, cargo generado y registrado como recibido.<br> N° Seguimiento:<b> $nu-$ano</b>");
-                            $r['e']=true;
-                        }else{
-                            if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddn;")){
-                                $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
-                            }else{
-                                $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
-                            }
-                        }
-
+                    $cnu=mysqli_query($cone, "SELECT idDoc FROM doc WHERE numdoc=$nu AND Ano='$ano';");
+                    if(mysqli_num_rows($cnu)>0){
+                        $r['m']=mensajewa("Error, # de seguimiento ya registrado. Vuelva a intentarlo para generar un uno nuevo.");
                     }else{
-                        $r['m']=mensajewa("Error al registrar cargo.");
-                        mysql_query($cone, "DELETE FROM doc WHERE idDoc=$iddn;");
-                    }            
+                        //ingresamos doc
+                        if(mysqli_query($cone, "INSERT INTO doc (Numero, Ano, Siglas, FechaDoc, idTipoDoc, folios, remitenteext, destinatarioext, deporigenext, depdestinoext, cargo, remitenteint, destinatarioint, deporigenint, depdestinoint, numdoc, asunto, fecregistro, regpor, idDocRel) VALUES ($numero, '$ano', $siglas, '$fechadoc', $idtipodoc, 1, $remitenteext, $destinatarioext, $deporigenext, $depdestinoext, 1, $remitenteint, $destinatarioint, $deporigenint, $depdestinoint, $nu, '$asunto', NOW(), $idem, $idd);")){
+                            $iddn=mysqli_insert_id($cone);
+
+                            //obtrenemos la mesa de partes
+                            $cmmp=mysqli_query($cone, "SELECT mp.idtdmesapartes FROM tdpersonalmp pm INNER JOIN tdmesapartes mp ON pm.idtdmesapartes=mp.idtdmesapartes WHERE pm.idEmpleado=$idem AND pm.estado=1 AND mp.estado=1;");
+                            if($rmmp=mysqli_fetch_assoc($cmmp)){
+                                $mmp=vacio($rmmp['idtdmesapartes']);
+                                $dep=vacio("");
+                            }else{
+                                $mmp=vacio("");
+                                $dep=vacio(iddependenciae($cone, $_SESSION['identi']));
+                            }
+                            mysqli_free_result($cmmp);
+
+                            $q="INSERT INTO tdestadodoc (idDoc, idtdestado, fecha, idEmpleado, idtdmesapartes, idDependencia, asignador, mpasignador, depasignador, estado) VALUES ($iddn, 2, NOW(), $idem, $mmp, $dep, $idem, $mmp, $dep, 1);";
+                            if(mysqli_query($cone, $q)){
+                                $r['m']=mensajesu("Listo, cargo generado y registrado como recibido.<br> N° Seguimiento:<b> $nu-$ano</b>");
+                                $r['e']=true;
+                            }else{
+                                if(mysqli_query($cone, "DELETE FROM doc WHERE idDoc=$iddn;")){
+                                    $r['m']=mensajewa("Error al registrar estado, vuelva a intentarlo. ".$q);
+                                }else{
+                                    $r['m']=mensajewa("Solo se registro el documento, contacte a informática para generarle un estado.");
+                                }
+                            }
+
+                        }else{
+                            $r['m']=mensajewa("Error al registrar cargo.");
+                            mysql_query($cone, "DELETE FROM doc WHERE idDoc=$iddn;");
+                        }           
+                    }
 
                 }else{
                     $r['m']=mensajewa('Error, datos erroneos del documento.');
